Relocate api,memberItem and DepartmentItem
